refactor: 폴더 구조 적용
