+++ calculator-app/build.gradle.kts
@@ -15,7 +15,6 @@ android {
         targetSdk = Version.targetSdk
         versionCode = 1
         versionName = "1.0"
-
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
     }
