--- calculator-domain/src/main/java/com/example/domain/OperationTerm.kt
@@ -1,3 +1,0 @@
-package com.example.domain
-
-interface OperationTerm
