+++ calculator-domain/src/main/java/com/example/domain/models/History.kt
@@ -1,3 +1,3 @@
-package com.example.domain
+package com.example.domain.models
 
 data class History(val statement: String, val result: Int)
