+++ calculator-domain/src/main/java/com/example/domain/models/OperationParser.kt
@@ -1,5 +1,4 @@
-package com.example.domain
-
+package com.example.domain.models
 
 object OperationParser {
