+++ calculator-domain/src/main/java/com/example/domain/models/OperationTerm.kt
@@ -1,0 +1,3 @@
+package com.example.domain.models
+
+interface OperationTerm
