+++ calculator-domain/src/main/java/com/example/domain/repositories/HistoryRepository.kt
@@ -1,6 +1,6 @@
 package com.example.domain.repositories
 
-import com.example.domain.History
+import com.example.domain.models.History
 
 interface HistoryRepository {
     suspend fun saveHistory(history: History): Unit
