+++ calculator-domain/src/test/java/com/example/domain/OperatorTest.kt
@@ -1,5 +1,6 @@
 package com.example.domain
 
+import com.example.domain.models.Operator
 import org.junit.Assert.*
 import org.junit.Test
 import org.junit.experimental.runners.Enclosed
