+++ calculator-domain/src/test/java/com/example/domain/StatementTest.kt
@@ -1,5 +1,6 @@
 package com.example.domain
 
+import com.example.domain.models.*
 import junit.framework.TestCase
 import org.junit.Assert.*
 import org.junit.Test
